Remove redundant HTML processors
The static HTML generated is now applied immediately to the page, instead of being re-sanitized.

I don't think we need to dynamically parse and sanitize it on the client-side, even if that means we can convert HTML `<a />` and `<img />` tags into NextJS `<Link />` and `<Image />` tags... Still! This is something to track and look into later, perhaps as a new low-priority issue.